RavenDB-12936 : counters work
